Large scale project restructuring
