+++ simon-gravelle/md_analysis_tutorial/polymer-in_water/data/raw/solvated_PEG.data
