+++ simon-gravelle/md_analysis_tutorial/polymer-in_water/src/constants/solvated_peg_constants.py
@@ -1,6 +1,6 @@
 class SolvatedPEGConstants:
     BASE_DIRECTORY = \
-        '/home/chris-vdb/Computational-Chemistry/lammps-tutorials/simon-gravelle/md_analysis_tutorial/python'
+        '/home/chris-vdb/Computational-Chemistry/lammps-tutorials/simon-gravelle/md_analysis_tutorial'
 
     PEG_MOLECULE_ATOMS = 'type 1 2 3 4 5 6 7'
