+++ simon-gravelle/md_analysis_tutorial/polymer-in_water/src/visualisation/solvated_peg_trajectory_plot.py
