+++ simon-gravelle/md_analysis_tutorial/src/constants/md_analysis_constants.py
