+++ simon-gravelle/md_analysis_tutorial/src/modules/scatter_plot.py
